+++ crates/rust-client/src/store/web_store/account/js_bindings.rs
@@ -29,6 +29,12 @@ extern "C" {
     #[wasm_bindgen(js_name = getAccountAssetVault)]
     pub fn idxdb_get_account_asset_vault(vault_root: String) -> js_sys::Promise;
 
+    #[wasm_bindgen(js_name = getAccountAuthByPubKey)]
+    pub fn idxdb_get_account_auth_by_pub_key(pub_key: String) -> JsValue;
+
+    #[wasm_bindgen(js_name = fetchAndCacheAccountAuthByPubKey)]
+    pub fn idxdb_fetch_and_cache_account_auth_by_pub_key(pub_key: String) -> js_sys::Promise;
+
     // INSERTS
     // ================================================================================================
 
@@ -56,6 +62,9 @@ extern "C" {
         account_seed: Option<Vec<u8>>,
         hash: String,
     ) -> js_sys::Promise;
+
+    #[wasm_bindgen(js_name = insertAccountAuth)]
+    pub fn idxdb_insert_account_auth(pub_key: String, secret_key: String) -> js_sys::Promise;
 
     #[wasm_bindgen(js_name = upsertForeignAccountCode)]
     pub fn idxdb_upsert_foreign_account_code(