+++ crates/web-client/src/account.rs
@@ -40,4 +40,45 @@ impl WebClient {
             Err(JsValue::from_str("Client not initialized"))
         }
     }
+
+    #[wasm_bindgen(js_name = "fetchAndCacheAccountAuthByAccountId")]
+    pub async fn fetch_and_cache_account_auth_by_account_id(
+        &mut self,
+        account_id: &AccountId,
+    ) -> Result<Option<String>, JsValue> {
+        let account_pub_key = {
+            let client = self
+                .get_mut_inner()
+                .ok_or_else(|| JsValue::from_str("Client not initialized"))?;
+
+            let account_record = client
+                .get_account(account_id.into())
+                .await
+                .map_err(|err| JsValue::from_str(&format!("Failed to get account: {err}")))?;
+
+            let account: NativeAccount = match account_record {
+                Some(record) => AccountRecord::into(record),
+                None => return Ok(None),
+            };
+
+            let account_storage = account.storage();
+            let pub_key_index = u8::from(account.is_faucet());
+
+            account_storage
+                .get_item(pub_key_index)
+                .map_err(|err| JsValue::from_str(&format!("Failed to get item: {err}")))?
+        };
+
+        let store =
+            self.store.as_ref().ok_or_else(|| JsValue::from_str("Store not initialized"))?;
+
+        let native_auth_secret_key = store
+            .fetch_and_cache_account_auth_by_pub_key(account_pub_key.to_hex())
+            .await
+            .map_err(|err| {
+                JsValue::from_str(&format!("Failed to fetch and cache account auth: {err}"))
+            })?;
+
+        Ok(native_auth_secret_key)
+    }
 }