+++ crates/web-client/src/new_account.rs
@@ -29,6 +29,7 @@ impl WebClient {
             keystore
                 .expect("KeyStore should be initialized")
                 .add_key(&AuthSecretKey::RpoFalcon512(key_pair))
+                .await
                 .map_err(|err| err.to_string())?;
 
             match client.add_account(&new_account, Some(account_seed), false).await {
@@ -91,6 +92,7 @@ impl WebClient {
             keystore
                 .expect("KeyStore should be initialized")
                 .add_key(&AuthSecretKey::RpoFalcon512(key_pair))
+                .await
                 .map_err(|err| err.to_string())?;
 
             match client.add_account(&new_account, Some(seed), false).await {