+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* Merge main into next and fix web keystore (#779).
 * Added support to import public accounts to `Client` (#733).
 * [BREAKING] Merged `TonicRpcClient` with `WebTonicRpcClient` and added missing endpoints (#744).
 
@@ -21,6 +22,19 @@
 ### Fixes
 
 * [BREAKING] Changed Snake Case Variables to Camel Case in JS/TS Files (#767).
+
+## 0.7.2 (2025-03-05) -  `miden-client-web` and `miden-client` crates
+
+### Changes
+
+* [BREAKING] Added initial Web Workers implementation to web client (#720, #743).
+* Web client: Exposed `InputNotes` iterator and `assets` getter (#757).
+* Web client: Exported `TransactionResult` in typings (#768).
+* Implemented serialization and deserialization for `SyncSummary` (#725).
+
+### Fixes
+
+* Web client: Fixed submit transaction; Typescript types now match underlying Client call (#760).
 
 ## 0.7.0 (2025-01-28)